Move preprocess.js back
